Remove namespaces TODO
It won't work as intended.

    %Rails:CustomTag

This code would have to invoke Rails.CustomTag, but CustomTag is an instance
method, not class method. If you make it a class method, you still have to worry
about getting the haml context from the caller, or creating a new haml context
in the instance. Creating a new haml context automatically makes sense for tags
defined in haml, but doesn't make sense for tags defined in ruby who want to
render children.